Display tab headers, allow profile to navigate again
